requirements.txt
Django==4.2
djangorestframework==3.14.0
requests~=2.28.2
djangorestframework-simplejwt==5.2.2
